add middleware

+++ app/Http/Controllers/UploadController.php
@@ -2,26 +2,50 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\videos;
 use Illuminate\Http\Request;
 use League\Flysystem\Filesystem;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Storage;
 use Hypweb\Flysystem\GoogleDrive\GoogleDriveAdapter;
 use Illuminate\Support\Facades\File;
+use FFMpeg;
+use Carbon\Carbon;
 
 class UploadController extends Controller
 {
-    //
-
-    public function testUpload(Request $request){
-        var_dump($request->all());
+    public function getUpload(){
+        return view('upload');
+    }
+    public function cdn_get($file_name){
+        return "https://anime47.imfast.io/$file_name";
+    }
+    public function getPreviewDrive($file_name){
+        $data = Storage::disk("google")->listContents();
+        foreach ($data as $file){
+            if ($file["name"] == $file_name){
+                return $file->path;
+            }
+        }
     }
 
     public function Upload(Request $request)
     {
+        var_dump($request->all());
 
         if ($request->hasFile('aksfileupload')) {
-            echo "Có";
+
+            /**
+             * Getting new id and user upload.
+             *
+             * @author Mai Viết Dũng.
+             */
+
+            $videos = new videos();
+            $last_video = $videos->getLast();
+            $new_id = $last_video + 1;
+            $new_format_id = sprintf("%02d", $new_id);
+            $user_id = $request->session()->get('LoggedUser')["id"];
 
             /**
              * Uploading to server.
@@ -30,8 +54,8 @@ class UploadController extends Controller
              */
 
             $file = $request->aksfileupload[0];
-            $original_file = $file->getClientOriginalName();
-            $file_edit_name = str_replace(" ", "_", $original_file);
+            $original_file = $file->getClientOriginalName()."_$new_format_id"."_$user_id";
+            $file_edit_name = str_replace(" ", "_", $original_file)."_$new_format_id"."_$user_id";
             $url = asset("temp_video").'/'.$file_edit_name;
             $file->move('temp_video', $file_edit_name);
 
@@ -44,15 +68,67 @@ class UploadController extends Controller
             $content = file_get_contents(asset($url));
             Storage::disk("google")->put($original_file, $content);
 
+
+
+            /**
+             * Get data to be inserted.
+             *
+             * @author
+             */
+            // Get the cdn.
+            $cdn_url = $this->cdn_get($original_file);
+
+            //getting preview and duration.
+            $ffmpeg = FFMpeg\FFMpeg::create(array( //config ffmpeg
+                'ffmpeg.binaries'  => '/usr/local/bin/ffmpeg',
+                'ffprobe.binaries' => '/usr/local/bin/ffprobe',
+                'timeout'          => 3600, // The timeout for the underlying process
+                'ffmpeg.threads'   => 12,   // The number of threads that FFMpeg should use
+            ));
+            $ffprobe = FFMpeg\FFProbe::create(array(
+                'ffmpeg.binaries'  => '/usr/local/bin/ffmpeg',
+                'ffprobe.binaries' => '/usr/local/bin/ffprobe',
+                'timeout'          => 3600, // The timeout for the underlying process
+                'ffmpeg.threads'   => 12,   // The number of threads that FFMpeg should use
+            ));
+
+            $get = $ffprobe->format($cdn_url)->get('duration',100);
+            $duration = gmdate("i:s", $get); // get the duration.
+
+            //get the preview.
+            $video = $ffmpeg->open(asset($url));
+            $frame = $video->frame(FFMpeg\Coordinate\TimeCode::fromSeconds(06));
+            $preview_file = $new_format_id."_screen.jpg"; //get the preview file name
+            mkdir("img/screens/$new_format_id", 0700);
+            $frame->save("img/screens/$new_format_id/$new_format_id"."_screen.jpg");
+
+            //get the upload date.
+            $time = Carbon::now();
+
+            // get drive url.
+            $drive_url = $this->getPreviewDrive($original_file);
+
+            /**
+             * Inserting to the database.
+             *
+             * @author
+             */
+
+            $insert = $videos->insert_video($preview_file, $cdn_url, $request->e3, $request->e4, $user_id
+            , $request->title, 0, $original_file, $new_format_id, $duration, $request->e2, $time, $drive_url
+
+            );
+
             /**
              * Empty the temp file.
              *
              * @author
              */
-            File::delete("temp_video/$file_edit_name");
 
-
-
+            if ($insert){
+                File::delete("temp_video/$file_edit_name");
+                return redirect()->route('home')->with(["success" => "Your video has been uploaded. Please wait for the system to process."]);
+            }
         }
     }
 }
